backend/requirements.txt
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn>=21.2.0
python-multipart>=0.0.18

# Gemini AI (new unified SDK)
# google-genai>=1.56.0 required by langchain-google-genai 4.x for Gemini 3 thought_signature support
google-genai>=1.56.0,<2.0.0

# Google Cloud Text-to-Speech — powers the daily audio coach brief
# (services/audio_coach.py). Lazy-imported; the service degrades to
# script-only if the dep is unavailable, so adding this is non-breaking.
google-cloud-texttospeech>=2.17.0,<3.0.0

# RAG & Vector Store
langchain>=0.3.0
# langchain-core>=1.2.5 required by langchain-google-genai 4.x
langchain-core>=1.2.5,<2.0.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
# langchain-google-genai 4.2.0+ required for Gemini 3 thought_signature support in tool calls
# Fixes: "400 Function call is missing a thought_signature in functionCall parts"
langchain-google-genai>=4.2.0,<5.0.0
# NOTE: langchain-google-vertexai removed - it pulls google-cloud-aiplatform (~200MB)
# which causes OOM on Render free tier (512MB). langchain-google-genai 4.x supports
# Vertex AI natively via vertexai=True parameter.
langgraph>=0.6.0

# Database
sqlalchemy==2.0.23
aiosqlite==0.19.0
supabase>=2.25.0
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis (shared cache across workers)
redis[hiredis]>=5.0.0

# Utilities
python-dotenv==1.0.0
pydantic>=2.7.4
pydantic-settings>=2.1.0
email-validator>=2.0.0

# HTTP Client
httpx>=0.28.0
websockets>=15.0

# Email
resend==2.0.0

# Firebase Admin SDK (for push notifications)
firebase-admin==6.5.0

# AWS Lambda
mangum==0.17.0
boto3==1.34.0

# Excel processing
pandas==2.1.4
openpyxl==3.1.2
pyarrow>=14.0.0

# Media / Video Processing
imageio-ffmpeg>=0.5.1

# Rate Limiting
slowapi==0.1.9

# Error tracking (Sentry) — silent no-op when SENTRY_DSN is unset
sentry-sdk[fastapi]>=2.0.0

# Testing
pytest>=8.2.0,<9.0.0
pytest-asyncio>=0.25.0,<1.0.0
pytest-cov>=5.0.0,<6.0.0
pytest-mock>=3.14.0,<4.0.0

# MCP (Model Context Protocol) server — publishes FitWiz tools to external AI clients
mcp>=1.2.0
# Report generation (PDF + HTML for generate_report tool)
weasyprint>=62.0
jinja2>=3.1.0
# Markdown → HTML conversion for MCP report generator
markdown>=3.5

# Gym Equipment Importer (AI importer feature)
# PDF page-count guard (native Gemini handles extraction; pypdf only counts pages)
pypdf>=4.0.0
# DOCX parsing for gym equipment / exercise imports
python-docx>=1.1.0
# HTML stripping for URL source (falls back to stdlib html.parser if import fails at runtime)
beautifulsoup4>=4.12.0

# ─── Workout-history / program-template import pipeline ─────────────────────
# Covers Hevy / Strong / Fitbod / Jefit / FitNotes / Garmin FIT / Apple Health /
# Strava / Peloton / Fitbit / Nike ingestion + Jeff Nippard / RP / Nuckols /
# Wendler / nSuns / GZCLP / Metallicadpa / Starting Strength creator programs,
# plus two-way OAuth sync + export back out (PDF / XLSX / GPX / TCX).
xlrd>=2.0.1                    # legacy .xls binary Excel (some older creator sheets)
chardet>=5.2.0                 # CSV encoding auto-detection (UTF-8/16, Windows-1252, ISO-8859-1)
dateparser>=1.2.0              # Hevy / year-less / locale date parsing
python-magic>=0.4.27           # MIME sniffing via libmagic (needs apt libmagic1 on Render)
lxml>=5.2.0                    # streaming Apple Health XML (500MB+ files)
fitparse>=1.2.0                # Garmin .FIT binary decoder
gpxpy>=1.6.2                   # Strava / Nike / Runkeeper GPX cardio files
stravalib>=1.7.0               # Strava OAuth + activity API
fitbit>=0.3.1                  # Fitbit Web API
garminconnect>=0.2.19          # Garmin Connect (consumer OAuth1; personal-use only)
reportlab>=4.2.0               # PDF export for the reverse-direction pipeline
polyline>=2.0.0                # Decode Google encoded polylines → GPX waypoints
cryptography>=42.0.0           # AES-GCM Fernet token encryption for oauth_sync_accounts
